requirements.txt
langchain
langchain-community
langchain-ollama
chromadb
streamlit
pymupdf
pypdf
sentence-transformers
numpy
pyyaml